Specify version property in allProjects in root project.
So. This was fun, and turned out to be variable scoping like I expected.

`version` is a project property set in each project. It was correctly set in every project. HOWEVER: I'm inferring from the behaviors of the forge prepareResources task that the forge plugin acts as its own subproject of the forge project, and so was passing a default version value.

allProjects runs for all projects, including sneaky little hidy projects we don't otherwise have direct access to, so this is the cleanest way to make sure the version property is everywhere.

I expect there's a way to bludgeon forge's little plugin into liking us, but I feel like this might be a more universal solution.