Improve error handling when self is the same as the base item
So if you search and replace badly you set the base for an item to itself and this renders as a StackOverflowException and a thousand line long stacktrace... and to add insult to injury it doesn't tell you where you made this boneheaded error.

Let me make your software more idiot-resistant.